+++ webext/background.js
@@ -316,8 +316,26 @@ async function onHeadersReceived_filterAltSvc(details) {
 
 async function onHeadersReceived_verifySelfAuthConnection(details) {
     let url = splitURL(details.url);
-    let secInfo = await browser.webRequest.getSecurityInfo(
-        details.requestId, {"certificateChain": true, "rawDER": true});
+    let secInfo = null;
+    try {
+      secInfo = await browser.webRequest.getSecurityInfo(
+          details.requestId, {"certificateChain": true, "rawDER": true});
+    } catch (e) {
+        if (e instanceof TypeError) {
+            log_debug(
+                "Caught exception trying to get securityInfo, assuming we",
+                "don't have that API in the browser:", e);
+            return;
+        } else {
+            throw e;
+        }
+    }
+
+    if (!secInfo) {
+        log_error("Failed to get securityInfo. No API support? Should have",
+            "caught this sooner.");
+        return;
+    }
 
     if (secInfo.state != "secure") {
         log_debug("Stopped considering", url.hostname, "becuase not",